Klargjort til x64 compilering. primært erstattet int/unsigned int med intptr_t/size_t som skifter til 64-bit ved x64 arkitektur Mangler at justere diverse exception-format-strenge så de kan håndtere int64 ved x64-compilering

+++ Lib/Include/Array.h
@@ -10,70 +10,70 @@ private:
   DECLARECLASSNAME;
   AbstractObjectManager *m_objectManager;
   void                 **m_elem;
-  unsigned long          m_size;
-  unsigned long          m_capacity;
-  unsigned long          m_updateCount;
+  size_t                 m_size;
+  size_t                 m_capacity;
+  size_t                 m_updateCount;
 
-  void resize();
-  void init(const AbstractObjectManager &objectManager, int size, unsigned long capacity);
-  int  getSortCount(int from, int count) const;
-  void sort(int from, int count, AbstractComparator &comparator);
-  int  linearSearch(const void *key, int from, int count, AbstractComparator &comparator) const;
-  int  binarySearch(const void *key, int from, int count, AbstractComparator &comparator) const;
-  int  binaryInsert(const void *e, AbstractComparator &comparator);
-  void indexError(const TCHAR *method, int index) const;
-  void selectError() const;
-  void unsupportedOperationError(const TCHAR *method) const;
+  void      resize();
+  void      init(const AbstractObjectManager &objectManager, size_t size, size_t capacity);
+  intptr_t  getSortCount(intptr_t from, intptr_t count) const;
+  void      sort(size_t from, size_t count, AbstractComparator &comparator);
+  intptr_t  linearSearch(const void *key, size_t from, size_t count, AbstractComparator &comparator) const;
+  intptr_t  binarySearch(const void *key, intptr_t from, size_t count, AbstractComparator &comparator) const;
+  intptr_t  binaryInsert(const void *e, AbstractComparator &comparator);
+  void      indexError(const TCHAR *method, size_t index) const;
+  void      selectError() const;
+  void      unsupportedOperationError(const TCHAR *method) const;
   friend class ArrayIterator;
 public:
-  ArrayImpl(AbstractObjectManager &objectManager, unsigned long capacity);
+  ArrayImpl(AbstractObjectManager &objectManager, size_t capacity);
   ArrayImpl(AbstractObjectManager &objectManager, const AbstractCollection *src);
   ~ArrayImpl();
   AbstractCollection *clone(bool cloneData) const;
   ArrayImpl &operator=(const ArrayImpl &src);
   bool add(const void *e);
-  bool add(unsigned int i, const void *e, unsigned int count);
-  void removeIndex(unsigned int i, unsigned int count);
+  bool add(size_t i, const void *e, size_t count);
+  void removeIndex(size_t i, size_t count);
   bool remove(const void *e);
   bool contains(const void *e) const; // unsuppoerted
   const void *select() const ;
   void *select();
-  inline int getCapacity() const {
+  inline size_t getCapacity() const {
     return m_capacity;
   }
-  void setCapacity(unsigned long capacity);
+  void setCapacity(size_t capacity);
   void clear();
-  inline int size() const {
+  inline size_t size() const {
     return m_size;
   }
-  inline void *getElement(unsigned int index) {
+  inline void *getElement(size_t index) {
     if(index >= m_size) indexError(_T("getElement"),index);
     return m_elem[index];
   }
 
-  inline const void *getElement(unsigned int index) const {
+  inline const void *getElement(size_t index) const {
     if(index >= m_size) indexError(_T("const getElement"),index);
     return m_elem[index];
   }
 
-  void swap(unsigned int i1, unsigned int i2);
-  void shuffle(int from, int count);
+  void swap(size_t i1, size_t i2);
+  void shuffle(size_t from, size_t count);
   bool equals(const ArrayImpl *rhs) const;
-  void arraySort(int from, int count, int (*compare)(const void **e1, const void **e2));
-  void arraySort(int from, int count, int (*compare)(const void  *e1, const void  *e2));
-  void arraySort(int from, int count, AbstractComparator &comparator);
+  void arraySort(size_t from, size_t count, int (*compare)(const void **e1, const void **e2));
+  void arraySort(size_t from, size_t count, int (*compare)(const void  *e1, const void  *e2));
+  void arraySort(size_t from, size_t count, AbstractComparator &comparator);
 
-  int lSearch(const void *key, int from, int count, int (*compare)(const void **e1, const void **e2)) const;
-  int lSearch(const void *key, int from, int count, int (*compare)(const void  *e1, const void  *e2)) const;
-  int lSearch(const void *key, int from, int count, AbstractComparator &comparator) const;
+  intptr_t lSearch(const void *key, size_t from, size_t count, int (*compare)(const void **e1, const void **e2)) const;
+  intptr_t lSearch(const void *key, size_t from, size_t count, int (*compare)(const void  *e1, const void  *e2)) const;
+  intptr_t lSearch(const void *key, size_t from, size_t count, AbstractComparator &comparator) const;
 
-  int bSearch(const void *key, int from, int count, int (*compare)(const void **e1, const void **e2)) const;
-  int bSearch(const void *key, int from, int count, int (*compare)(const void  *e1, const void  *e2)) const;
-  int bSearch(const void *key, int from, int count, AbstractComparator &comparator) const;
+  intptr_t bSearch(const void *key, size_t from, size_t count, int (*compare)(const void **e1, const void **e2)) const;
+  intptr_t bSearch(const void *key, size_t from, size_t count, int (*compare)(const void  *e1, const void  *e2)) const;
+  intptr_t bSearch(const void *key, size_t from, size_t count, AbstractComparator &comparator) const;
 
-  int bInsert(const void *e, int (*compare)(const void **e1, const void **e2));
-  int bInsert(const void *e, int (*compare)(const void  *e1, const void  *e2));
-  int bInsert(const void *e, AbstractComparator &comparator);
+  intptr_t bInsert(const void *e, int (*compare)(const void **e1, const void **e2));
+  intptr_t bInsert(const void *e, int (*compare)(const void  *e1, const void  *e2));
+  intptr_t bInsert(const void *e, AbstractComparator &comparator);
 
   AbstractIterator *getIterator();
 };
@@ -83,7 +83,7 @@ public:
   Array() : Collection<T>(new ArrayImpl(ObjectManager<T>(), 10)) {
   }
 
-  explicit Array(unsigned long capacity) : Collection<T>(new ArrayImpl(ObjectManager<T>(), capacity)) {
+  explicit Array(size_t capacity) : Collection<T>(new ArrayImpl(ObjectManager<T>(), capacity)) {
   }
 
   Array(const Collection<T> &src) : Collection<T>(new ArrayImpl(ObjectManager<T>(), src.size())) {
@@ -99,11 +99,11 @@ public:
     return *this;
   }
 
-  inline T &operator[](unsigned int i) {
+  inline T &operator[](size_t i) {
     return *(T*)(((ArrayImpl*)m_collection)->getElement(i));
   }
   
-  inline const T &operator[](unsigned int i) const {
+  inline const T &operator[](size_t i) const {
     return *(T*)(((ArrayImpl*)m_collection)->getElement(i));
   }
   
@@ -124,11 +124,11 @@ public:
   }
   
   bool operator==(const Array<T> &a) const { // NB not virtual in Collection, because of ==
-    const int n = size();
+    const size_t n = size();
     if(n != a.size()) {
       return false;
     }
-    for(int i = 0; i < n; i++) {
+    for(size_t i = 0; i < n; i++) {
       if(!((*this)[i] == a[i])) {
         return false;
       }
@@ -140,7 +140,7 @@ public:
     return !(operator==(a));
   }
 
-  inline bool add(unsigned int i, const T &e, unsigned int count=1) {
+  inline bool add(size_t i, const T &e, size_t count=1) {
     return ((ArrayImpl*)m_collection)->add(i, &e, count);
   }
 
@@ -149,8 +149,8 @@ public:
   }
 
   inline bool remove(const T &e) {                    // NB not virtual in Collection, because of ==
-    const int n = size();
-    for(int i = 0; i < n; i++) {
+    const size_t n = size();
+    for(size_t i = 0; i < n; i++) {
       if((*this)[i] == e) {
         removeIndex(i);
         return true;
@@ -159,7 +159,7 @@ public:
     return false;
   }
 
-  inline bool removeIndex(unsigned int i, unsigned int count=1) {
+  inline bool removeIndex(size_t i, size_t count=1) {
     ((ArrayImpl*)m_collection)->removeIndex(i, count);
     return count > 0;
   }
@@ -172,9 +172,9 @@ public:
     return getFirstIndex(e) >= 0;
   }
 
-  int getFirstIndex(const T &e) const {
-    const int n = size();
-    for(int i = 0; i < n; i++) {
+  intptr_t getFirstIndex(const T &e) const {
+    const size_t n = size();
+    for(size_t i = 0; i < n; i++) {
       if((*this)[i] == e) {
         return i;
       }
@@ -182,12 +182,12 @@ public:
     return -1;
   }
 
-  inline Array<T> &swap(unsigned int i1, unsigned int i2) {
+  inline Array<T> &swap(size_t i1, size_t i2) {
     ((ArrayImpl*)m_collection)->swap(i1, i2);
     return *this;
   }
 
-  inline Array<T> &shuffle(int from, int count) {
+  inline Array<T> &shuffle(size_t from, size_t count) {
     ((ArrayImpl*)m_collection)->shuffle(from, count);
     return *this;
   }
@@ -202,7 +202,7 @@ public:
   };
 
 private:
-  bool permuter(int n, PermutationHandler &handler) {
+  bool permuter(size_t n, PermutationHandler &handler) {
     if(n <= 1) {
       return handler.handlePermutation(*this);
     } else {
@@ -210,7 +210,7 @@ private:
       if(!permuter(n, handler)) {
         return false;
       }
-      for(int i = 0; i < n; i++) {
+      for(size_t i = 0; i < n; i++) {
         swap(i, n);
         if(!permuter(n, handler)) {
           return false;
@@ -227,7 +227,7 @@ public:
     return permuter(size(), handler);
   }
 
-  inline Array<T> &sort(int from, int count, int (*compare)(const T **e1, const T **e2)) {
+  inline Array<T> &sort(size_t from, size_t count, int (*compare)(const T **e1, const T **e2)) {
     ((ArrayImpl*)m_collection)->arraySort(from, count, (int(*)(const void **, const void **))compare);
     return *this;
   }
@@ -236,7 +236,7 @@ public:
     return sort(0, size(), compare);
   }
 
-  Array<T> &sort(int from, int count, int (*compare)(const T &e1 , const T &e2)) {
+  Array<T> &sort(size_t from, size_t count, int (*compare)(const T &e1 , const T &e2)) {
     ((ArrayImpl*)m_collection)->arraySort(from, count, (int(*)(const void *, const void *))compare);
     return *this;
   }
@@ -245,7 +245,7 @@ public:
     return sort(0, size(), compare);
   }
 
-  Array<T> &sort(int from, int count, Comparator<T> &comparator) {
+  Array<T> &sort(size_t from, size_t count, Comparator<T> &comparator) {
     ((ArrayImpl*)m_collection)->arraySort(from, count, comparator);
     return *this;
   }
@@ -254,75 +254,75 @@ public:
     return sort(0, size(), comparator);
   }
 
-  int search(const T &key, int from, int count, int (*compare)(const T **e1, const T **e2)) const {
+  intptr_t search(const T &key, size_t from, size_t count, int (*compare)(const T **e1, const T **e2)) const {
     return ((ArrayImpl*)m_collection)->lSearch(&key, from, count, (int(*)(const void **, const void **))compare);
   }
 
-  int search(const T &key, int (*compare)(const T **e1, const T **e2)) const {
+  intptr_t search(const T &key, int (*compare)(const T **e1, const T **e2)) const {
     return search(key, 0, size(), compare);
   }
 
-  int search(const T &key, int from, int count, int (*compare)(const T &e1, const T &e2)) const {
+  intptr_t search(const T &key, size_t from, size_t count, int (*compare)(const T &e1, const T &e2)) const {
     return ((ArrayImpl*)m_collection)->lSearch(&key, from, count, (int(*)(const void *, const void *))compare);
   }
 
-  int search(const T &key, int (*compare)(const T &e1, const T &e2)) const {
+  intptr_t search(const T &key, int (*compare)(const T &e1, const T &e2)) const {
     return search(key, 0, size(), compare);
   }
 
-  int search(const T &key, int from, int count, Comparator<T> &comparator) const {
+  intptr_t search(const T &key, size_t from, size_t count, Comparator<T> &comparator) const {
     return ((ArrayImpl*)m_collection)->lSearch(&key, from, count, comparator);
   }
 
-  int search(const T &key, Comparator<T> &comparator) const {
+  intptr_t search(const T &key, Comparator<T> &comparator) const {
     return search(key, 0, size(), comparator);
   }
 
-  int binarySearch(const T &key, int from, int count, int (*compare)(const T **e1, const T **e2)) const {
+  intptr_t binarySearch(const T &key, size_t from, size_t count, int (*compare)(const T **e1, const T **e2)) const {
     return ((ArrayImpl*)m_collection)->bSearch(&key, from, count, (int(*)(const void **, const void **))compare);
   }
 
-  int binarySearch(const T &key, int (*compare)(const T **e1, const T **e2)) const {
+  intptr_t binarySearch(const T &key, int (*compare)(const T **e1, const T **e2)) const {
     return binarySearch(key, 0, size(), compare);
   }
 
-  int binarySearch(const T &key, int from, int count, int (*compare)(const T &e1, const T &e2)) const {
+  intptr_t binarySearch(const T &key, size_t from, size_t count, int (*compare)(const T &e1, const T &e2)) const {
     return ((ArrayImpl*)m_collection)->bSearch(&key, from, count, (int(*)(const void *, const void *))compare);
   }
 
-  int binarySearch(const T &key, int (*compare)(const T &e1, const T &e2)) const {
+  intptr_t binarySearch(const T &key, int (*compare)(const T &e1, const T &e2)) const {
     return binarySearch(key, 0, size(), compare);
   }
 
-  int binarySearch(const T &key, int from, int count, Comparator<T> &comparator) const {
+  intptr_t binarySearch(const T &key, size_t from, size_t count, Comparator<T> &comparator) const {
     return ((ArrayImpl*)m_collection)->bSearch(&key, from, count, comparator);
   }
 
-  int binarySearch(const T &key, Comparator<T> &comparator) const {
+  intptr_t binarySearch(const T &key, Comparator<T> &comparator) const {
     return binarySearch(key, 0, size(), comparator);
   }
 
-  int binaryInsert(const T &key, int (*compare)(const T **e1, const T **e2)) {
+  intptr_t binaryInsert(const T &key, int (*compare)(const T **e1, const T **e2)) {
     return ((ArrayImpl*)m_collection)->bInsert(&key, (int(*)(const void **, const void **))compare);
   }
 
-  int binaryInsert(const T &key, int (*compare)(const T  *e1, const T  *e2)) {
+  intptr_t binaryInsert(const T &key, int (*compare)(const T  *e1, const T  *e2)) {
     return ((ArrayImpl*)m_collection)->bInsert(&key, (int(*)(const void *, const void *))compare);
   }
 
-  int binaryInsert(const T &key, int (*compare)(const T &e1, const T &e2)) {
+  intptr_t binaryInsert(const T &key, int (*compare)(const T &e1, const T &e2)) {
     return ((ArrayImpl*)m_collection)->bInsert(&key, (int(*)(const void *, const void *))compare);
   }
 
-  int binaryInsert(const T &key, Comparator<T> &comparator) {
+  intptr_t binaryInsert(const T &key, Comparator<T> &comparator) {
     return ((ArrayImpl*)m_collection)->bInsert(&key, comparator);
   }
 
-  inline int getCapacity() const {
+  inline size_t getCapacity() const {
     return ((ArrayImpl*)m_collection)->getCapacity();
   }
 
-  inline void setCapacity(int capacity) {
+  inline void setCapacity(size_t capacity) {
     ((ArrayImpl*)m_collection)->setCapacity(capacity);
   }
 };
@@ -334,8 +334,8 @@ public:
   StringArray() {}
   StringArray(const char **strArray); // terminate with NULL-pointer
   StringArray(Tokenizer &tok);
-  int maxLength() const;
-  int minLength() const;
+  size_t maxLength() const;
+  size_t minLength() const;
   String getAsDoubleNullTerminatedString() const; // return a double 0-terminated string containing all strings of the array, each 0-terminated
 };
