+++ Lib/Include/ByteStream.h
@@ -4,14 +4,14 @@
 
 class ByteInputStream {
 public:
-  virtual int getBytes(BYTE *dst, unsigned int n) = 0; // Must return the actual number of bytes read.
+  virtual intptr_t getBytes(BYTE *dst, size_t n) = 0; // Must return the actual number of bytes read.
   virtual int getByte() = 0;                           // Must return EOF when end-of-stream
-  void getBytesForced(BYTE *dst, unsigned int n);      // Throws "unexpected end-of-stream Exception", if the requested number of bytes cannot be read
+  void getBytesForced(BYTE *dst, size_t n);      // Throws "unexpected end-of-stream Exception", if the requested number of bytes cannot be read
 };
 
 class ByteOutputStream {
 public:
-  virtual void putBytes(const BYTE *src, unsigned int n) = 0;
+  virtual void putBytes(const BYTE *src, size_t n) = 0;
   virtual void putByte(BYTE b) = 0;
 };
